requirements.txt
pandas
scikit-learn
seaborn
matplotlib.pyplot
google-cloud-bigquery
fuzzywuzzy
python-Levenshtein
IPython
category_encoders
tensorflow